Replace category and tag objects with simple string (#6)
the API doesn't document possible `id`s or `name`s for either, and doesn't provide endpoints to look them up so it's not clear how one producing or consuming an implementation would know what to do with them. 

Also it doesn't make sense to send both the id and name in a request.